Add extend_gg_call call to add extra features to the esquisse provided ggplot2 metaprogramming. Its arguments are passed to esquisse-server via ...
For now, it just sets the scale of the x axis depending on the t_unit argument

This argument is shared with t_unit in controls_server, which uses it to generate an extra filtering input for time with the corresponding unit

safe_ggplot needs to be modified so the eval_tidy call uses the parent environment. This is so because extend_gg_call has the side effect of defining new names which and exporting them to its parent frame, which is also safe_ggplot's parent.frame, ergo using the parent.frame makes them visible to safe_ggplot.